Enhance user interface and functionality with profile picture support, improved form modals, and search features

+++ front-end/js/main.js
@@ -19,10 +19,10 @@ import {
 import {
   createMap,
   favoriteMarker,
-  getRestaurantsLatLon,
+  getRestaurantsLatLon, userMarker,
   zoomRestaurant
 } from "./map.js"
-import {baseUrl, restaurantUrl,} from "./variables.js";
+import {baseUrl, restaurantUrl, uploadsUrl,} from "./variables.js";
 
 let restaurants = [];
 let favorites = []
@@ -35,7 +35,9 @@ const div = document.querySelector('.restaurant-container')
 const loginModal = document.querySelector('dialog');
 const signUpModal = document.querySelector('dialog');
 const userMenuModal = document.querySelector('dialog');
-const menuDisplay = document.querySelector('.menu-display')
+const menuDisplay = document.querySelector('.menu-display');
+const searchform = document.querySelector('.search-form');
+let box;
 
 //Getters for restaurant API's
 const getRestaurants = async () => {
@@ -143,7 +145,7 @@ const getFavoritesByUserId = async () => {
   }
 }
 
-const getUserLocation =  () => {
+const getUserLocation = () => {
   return new Promise((resolve, reject) => {
     navigator.geolocation.getCurrentPosition(
       (position) => resolve(position),
@@ -158,7 +160,7 @@ const calculateDistance = (lat1, lon1, lat2, lon2) => {
 
 const findNearestRestaurant = async (restaurants) => {
   const user = await getUserLocation();
-  console.log(user)
+  userMarker(user);
   let closestRestaurant = null;
   let minDistance = Infinity;
   for (let rest of restaurants) {
@@ -172,8 +174,11 @@ const findNearestRestaurant = async (restaurants) => {
       closestRestaurant = rest;
     }
   }
-  console.log(closestRestaurant)
   zoomRestaurant(closestRestaurant);
+  const menu = await getDailyMenu(closestRestaurant._id, 'en');
+  menuDisplay.innerHTML = menuTable(menu);
+  showWeeklyMenu(closestRestaurant)
+
 };
 
 const signUpPage = () => {
@@ -182,13 +187,26 @@ const signUpPage = () => {
     signUpModal.innerHTML = signUpScreen();
     signUpModal.showModal();
 
+    const closeBtn = document.querySelector('.close-btn')
+    closeBtn.addEventListener('click', function () {
+      signUpModal.close();
+    })
+
     requestAnimationFrame(() => {
       const signUpForm = document.querySelector('#signUpForm');
-      signUpForm.addEventListener('submit', function (e) {
+      signUpForm.addEventListener('submit', async function (e) {
         e.preventDefault();
         const formData = new FormData(signUpForm);
         console.log(formData)
-        registerUser(formData);
+        await registerUser(formData);
+
+        const username = formData.get('username');
+        const password = formData.get('password');
+
+        if (username && password) {
+          const credentials = {username, password};
+          await postUserCredentials(credentials);
+        }
       });
     });
   });
@@ -197,14 +215,21 @@ const signUpPage = () => {
 
 const updateMenu = () => {
   if (currentUser) {
+    console.log(currentUser.profile_picture)
+    console.log(`${uploadsUrl}${currentUser.profile_picture}_thumb.png`);
     headersUl.innerHTML = `
+    <img class="profile-pic" src="${uploadsUrl}${currentUser.profile_picture}_thumb.png" alt="profile-pic">
     <li id="userMenuBtn">User Menu</li>
     <li id="logoutBtn">Logout</li>
     `;
     document.querySelector('#userMenuBtn').addEventListener('click', () => {
-      userMenuModal.innerHTML = userMenuScreen();
+      userMenuModal.innerHTML = userMenuScreen(currentUser);
       userMenuModal.showModal();
       userMenuPage();
+      const closeBtn = document.querySelector('.close-btn')
+      closeBtn.addEventListener('click', function (e) {
+        userMenuModal.close();
+      })
 
     });
     document.querySelector('#logoutBtn').addEventListener('click', () => {
@@ -218,35 +243,41 @@ const updateMenu = () => {
 
 
 const userMenuPage = () => {
-  const userMenuForm = document.querySelector('#userMenuForm');
-  userMenuForm.addEventListener('submit', function (e) {
-    e.preventDefault();
-    const clickedButton = e.submitter;
-    const action = clickedButton.value;
-    console.log(action)
-    if (action === 'update') {
-      const formData = new FormData(userMenuForm);
-      const userData = {};
-      formData.forEach((value, key) => {
-        if (!(value instanceof File)) {
-          userData[key] = value;
+  if (currentUser) {
+    const userMenuForm = document.querySelector('#userMenuForm');
+    userMenuForm.addEventListener('submit', function (e) {
+      e.preventDefault();
+      const clickedButton = e.submitter;
+      const action = clickedButton.value;
+      console.log(action);
+      if (action === 'update') {
+        const formData = new FormData(userMenuForm);
+        const userData = {};
+        formData.forEach((value, key) => {
+          if (!(value instanceof File) && value.trim() !== '') {
+            userData[key] = value;
+          }
+        });
+        if (Object.keys(userData).length > 0) {
+          putUser(userData);
         }
-      });
-      putUser(userData);
-    }
-    if (action === 'delete') {
-      deleteUser();
-    }
-  });
+      }
+      if (action === 'delete') {
+        deleteUser();
+      }
+    });
+  }
 };
-
 
 const loginPage = () => {
   const loginBtn = document.querySelector('#loginBtn');
   loginBtn.addEventListener('click', function () {
     loginModal.innerHTML = loginScreen();
     loginModal.showModal();
-
+    const closeBtn = document.querySelector('.close-btn')
+    closeBtn.addEventListener('click', function (e) {
+      loginModal.close();
+    })
     const loginForm = document.querySelector('#loginForm');
     loginForm.addEventListener('submit', function (e) {
       e.preventDefault();
@@ -262,12 +293,16 @@ const loginPage = () => {
 }
 
 
-const createContainerBoxes = async () => {
+const createContainerBoxes = async (restaurantList) => {
   div.innerHTML = '';
-  findNearestRestaurant(restaurants);
-  for (const restaurant of restaurants) {
+  console.log('pyöriii')
+  findNearestRestaurant(restaurants)
+  if (!restaurantList) {
+    restaurantList = restaurants;
+  }
+  for (const restaurant of restaurantList) {
     try {
-      const box = restaurantBox(restaurant);
+      box = restaurantBox(restaurant);
       div.append(box);
       const menu = await getDailyMenu(restaurant._id, 'en');
       const menuHtml = restaurantMenu(menu);
@@ -276,10 +311,11 @@ const createContainerBoxes = async () => {
       if (Array.isArray(favorites) && favorites.some(fav => Number(fav.company_id) === restaurant.companyId)) {
         heartBtn.classList.add('favorite');
       }
-      box.addEventListener('click', async function() {
+      box.addEventListener('click', async function () {
         try {
           const menu = await getDailyMenu(restaurant._id, 'en');
           menuDisplay.innerHTML = menuTable(menu);
+          zoomRestaurant(restaurant);
           showWeeklyMenu(restaurant)
         } catch (error) {
           console.log(error);
@@ -310,6 +346,31 @@ const createContainerBoxes = async () => {
 };
 
 
+searchform.addEventListener('submit', function (e) {
+  e.preventDefault();
+  const searchValue = searchform.searchvalue.value.trim().toLowerCase();
+
+
+  const filteredRestaurants = restaurants.filter(restaurant =>
+    restaurant.name.toLowerCase().includes(searchValue)
+  );
+  console.log(searchValue);
+  createContainerBoxes(filteredRestaurants);
+});
+
+const favButton = document.querySelector('#fav-button');
+
+favButton.addEventListener('click', function () {
+  if (!Array.isArray(favorites)) return;
+
+  const favoriteRestaurants = restaurants.filter(restaurant =>
+    favorites.some(fav => Number(fav.company_id) === restaurant.companyId)
+  );
+
+  createContainerBoxes(favoriteRestaurants);
+});
+
+
 const showWeeklyMenu = (restaurant) => {
   const dayButtons = document.querySelectorAll('button[data-day]');
   dayButtons.forEach(button => {
@@ -318,14 +379,14 @@ const showWeeklyMenu = (restaurant) => {
     newButton.addEventListener('click', async (e) => {
       e.preventDefault();
       const day = e.target.dataset.day;
-        try {
-          weeklyMenu = [];
-          await getWeeklyMenu(restaurant._id, 'en');
-          const selectedDayMenu = weeklyMenu.days[day];
-          menuDisplay.innerHTML = menuTable(selectedDayMenu);
-        } catch (error) {
-          console.log(error);
-        }
+      try {
+        weeklyMenu = [];
+        await getWeeklyMenu(restaurant._id, 'en');
+        const selectedDayMenu = weeklyMenu.days[day];
+        menuDisplay.innerHTML = menuTable(selectedDayMenu);
+      } catch (error) {
+        console.log(error);
+      }
 
     });
   });
@@ -351,6 +412,7 @@ const showNotification = (targetElement, message, timeout = 3000) => {
 
 const init = async () => {
   await getRestaurants();
+  await createMap();
   await getRestaurantsLatLon(restaurants);
   createContainerBoxes();
   loginPage();
